Integration Tests
https://eaflood.atlassian.net/jira/software/projects/FUT/boards/1668?selectedIssue=FUT-344

Move testcontainers config into seperate setup module
