+++ test/grant-api.test.js
@@ -1,6 +1,5 @@
 import { describe, it, before, after, beforeEach } from "node:test";
 import { assert } from "../src/common/assert.js";
-import { DockerComposeEnvironment, Wait } from "testcontainers";
 import Wreck from "@hapi/wreck";
 import { MongoClient } from "mongodb";
 import { config } from "../src/common/config.js";
@@ -8,39 +7,17 @@ import { grant1, grant2 } from "./fixtures/grants.js";
 import { collection as grantsCollection } from "../src/grants/grant-repository.js";
 
 describe("Grant API Tests", () => {
-  const PORT = 3001;
-  const MONGO_PORT = 28017;
-  let appUrl;
-  let environment;
-  let mongoUri;
   let db;
   let client;
 
   before(async () => {
-    environment = await new DockerComposeEnvironment(".", "compose.yml")
-      .withEnvironment({ PORT: PORT.toString() })
-      .withEnvironment({ MONGO_PORT: MONGO_PORT.toString() })
-      .withWaitStrategy("mongodb", Wait.forListeningPorts())
-      .withWaitStrategy("gas", Wait.forLogMessage("server started"))
-      .up();
-
-    // Get MongoDB container URI
-    const container = environment.getContainer("mongodb-1");
-    const mongoPort = container.getMappedPort(MONGO_PORT);
-    const host = container.getHost();
-    mongoUri = `mongodb://${host}:${mongoPort}`;
-    appUrl = `http://${host}:${PORT}`;
-
-    // Connect to MongoDB
-    client = new MongoClient(mongoUri);
+    client = new MongoClient(global.MONGO_URI);
     await client.connect();
     db = client.db(config.get("mongoDatabase"));
   });
 
   after(async () => {
-    // Shutdown and cleanup the environment
     if (client) await client.close();
-    if (environment) await environment.stop();
   });
 
   describe("POST /grants", () => {
@@ -49,8 +26,7 @@ describe("Grant API Tests", () => {
     });
 
     it("add a grant", async () => {
-      // Make a POST request to add the grant
-      const postResponse = await Wreck.post(`${appUrl}/grants`, {
+      const postResponse = await Wreck.post(`${global.APP_URL}/grants`, {
         json: true,
         payload: grant1,
       });
@@ -73,7 +49,9 @@ describe("Grant API Tests", () => {
     it("Find grants", async () => {
       await db.collection(grantsCollection).insertMany([grant1, grant2]);
 
-      const getResponse = await Wreck.get(`${appUrl}/grants`, { json: true });
+      const getResponse = await Wreck.get(`${global.APP_URL}/grants`, {
+        json: true,
+      });
       assert.equal(getResponse.res.statusCode, 200);
       assert.equal(getResponse.payload.length, 2);
       assert.equal(getResponse.payload[0].code, "e2e-code1");
@@ -89,9 +67,12 @@ describe("Grant API Tests", () => {
     it("Find grant by code", async () => {
       await db.collection(grantsCollection).insertMany([grant1, grant2]);
 
-      const getResponse = await Wreck.get(`${appUrl}/grants/e2e-code2`, {
-        json: true,
-      });
+      const getResponse = await Wreck.get(
+        `${global.APP_URL}/grants/e2e-code2`,
+        {
+          json: true,
+        },
+      );
       assert.equal(getResponse.res.statusCode, 200);
       assert.equal(getResponse.payload.code, "e2e-code2");
     });
